requirements.txt
aiohttp==1.0.5
async-timeout==2.0.0
chardet==3.0.4
discord.py==0.16.12
multidict==3.3.0
pkg-resources==0.0.0
websockets==3.4
